sigo sin tener idea de como funciona la vida en nuestro juego y freeze frames corregidos

+++ SystemTest/Assets/Mock-Up/Part.cs
@@ -6,7 +6,7 @@ public class Part : MonoBehaviour
 {
     public string PartName;
     public string AttName;
-    public int life;
+    public float life;
     public int Damage;
     public float Aspeed;
     public AudioClip AttackSound;
